Use custom package name for sample.
This commit adds a suffix (.debug) to the package name of the
sample watch face when built in the debug configuration. This is
done so that the sample can be installed on the emulator for
API 30 and above, where it is pre-installed by default, which
would other lead to a conflict in the signing keys as a developer
trying out the sample will be using a different signing key.

Keep the release configuration unchanged as developers do not
need to use it and it is can instead be used to generate the
right APK to be included in the emulator image.

Bug: 192636368
Test: Verifying that ./gradlew :wear:wear-watchface-samples-app:installDebug works
Change-Id: I143a72b5f0d667dea3d04e96351ebe2744b6394e