Skipping trying to run lint in com.android.test projects
They don't have any lint tasks.

This will allow us to switch from lintDebug to lintRelease.
:benchmark:integration-tests:test-module-sample disables its 'debug' variant, so
if we try to switch from lintDebug to lintRelease then we fail to find the task in this project,
which only uses the com.android.test plugin and not com.android.library or com.android.application.

Bug: 186569865
Test: Treehugger runs busytown/androidx.sh
Test: Run `gw bOS -m | grep -i lint` before and after this change and see that its output is unchanged

Change-Id: Ie5b42651d7a32af3e88fa2a2adfa024ffa61d9a7