Allows ImageAnalysis to select a resolution larger than 1080p
Remove the 1080p max resolution setting from ImageAnalysis.

- A LEGACY-level device can still only support PRIV/PREVIEW + YUV/PREVIEW + JPEG/MAXIMUM configuration for Preview + ImageAnalysis + ImageCapture. A resolution larger than 1080p can't be selected is due to the device capability but not the restriction of CameraX.
- LIMITED-level above devices can support PRIV/PREVIEW + YUV/RECORD + JPEG/RECORD configuration for Preview + ImageAnalysis + ImageCapture. Since this configuration is listed in the guaranteed supported configurations tables, it should be safe to use a larger resolution for ImageAnalysis.
- The default 640x480 resolution setting of ImageAnalysis is not changed. This CL should not impact the selected resolutions of all use cases when the application doesn't explicitly specify they want a larger resolution for ImageAnalysis.
- The selected resolution of ImageCapture will only be affected when the application explicitly specifies that it wants a larger resolution for ImageAnalysis rather than a MAXIMUM resolution for ImageCapture.

Relnote: "Allows ImageAnalysis to select a resolution larger than 1080p. A LIMITED-level above device can support a RECORD size resolution for ImageAnalysis when it is bound together with Preview and ImageCapture. The trade-off is the selected resolution for the ImageCapture will also need to be a RECORD size resolution. To successfully select a RECORD size resolution for ImageAnalysis, a RECORD size target resolution should be set on both ImageCapture and ImageAnalysis. This indicates that the application clearly understand the trade-off and prefer the ImageAnalysis to have a larger resolution rather than the ImageCapture to have a MAXIMUM resolution. For the definitions of RECORD, MAXIMUM sizes and more details see https://developer.android.com/reference/android/hardware/camera2/CameraDevice#regular-capture. The RECORD size refers to the camera device's maximum supported recording resolution, as determined by CamcorderProfile. The MAXIMUM size refers to the camera device's maximum output resolution for that format or target from StreamConfigurationMap.getOutputSizes(int)."

Bug: 192911449
Test: ./gradlew bOS
Change-Id: I1ee973a50f9605324f06f406ca2230a464ed2363