Add support for FilterQuality in DrawScope
Relnote: "Updated DrawScope#drawImage method that
consumes source and destination rects
to consume an optional FilterQuality
parameter. This is useful for pixel
art that is intended to be pixelated
when scaled up for pixel based art.
Updated BitmapPainter + Image composable
to also consume an optional FilterQuality
parameter"

Fixes: 180311607
Test: Added various tests to DrawScopeTest,
BitmapPainterTest and ImageTest

Change-Id: Ie4fb04013701add0fba1c5c6bb9da2812d6436e7