Only draw a blank invisible frame for below R.
In previous versions of Wear OS (below R), it was necessary for the
system to draw a blank frame before the watch face became invisible,
to help with the transition between different modes.

This is no longer needed on the latest version of Wear OS, so we
should remove this from the library.

Bug: 193110049
Bug: 193089651
Test: manually on device
Change-Id: Idfbfe68111d2e255adf05033c6dc2006e5d9a986