Adds in-app-memory tracing, for tracing out-of-scope macrobench work
Bug:194424816
Test: ./gradlew benchmark:benchmark-macro:cC
Test: TrivialStartupTest
Relnote: N/A

It's userspace tracing! ...sort of...

This in-app-memory tracing can capture sections outside of the main
measure block, without impacting metrics - they're added after metric
processing.

This is done by manually creating perfetto protos in a list, and
appending a trace containing these events to the end of the final
trace file, after metrics are extracted, similar to UiState.

The downside to this approach is that unterminated trace events appear
to last > 1 second, so the behavior is disabled currently.

Adds tracing throughout macrobenchmark lifecycle for tasks such as
perfetto start/stop, and metric extraction.

Extends manually-pruned protos to include newly required sections.

Change-Id: I0e10208f9eccf938343333f723238ec4135b60a5