Use java 8 classes for times in watch face public APIs
The underlying wireformat is unchanged so this change is
backwards compatiable. We also add equals and hashCode method to
PreviewScreenshotParams. Apart from that this is a pure refactor
with no functional changes.

Test: Presubmit
Bug: 194230792
Relnote: All public watch face, client, editor and complication APIs now use java.time.Instant for times rather than a Long, as a consquence the minimum API level has increaced to 26.
Change-Id: I3cd482ea3c80e6b88d28f9e33368b8f0e8e651c8