Add equals and hashcode to styling types.
These types are used in maps and can be tested for equality, so it is
important that we define equals and hashcode methods for them so that
they can be used correctly.

This will cause issues if multiple instances of the same type of item
are created, e.g., if the watch face and the editor have different
instances. This is likely as a context is required to create these
objects and the watch face and editor do not share the same context.

Bug: 194476557
Test: unit tests
Change-Id: I67c1c737472ae3c090779ed18a06d895c238600d