Add minimum touch target size for Semantics and Pointer Input
Bug: 189106692

Minimum touch target size has been added to ViewConfiguration
and used in pointer input and accessibility onClick() to
ensure that clickable containers are accessible.

When a touch is not a direct hit, the nearest target that
hits in the minimum touch target region is considered hit.
This will affect hit testing performance to some extent, but
shouldn't be an order of magnitude different.

RelNote: "Added minimum touch target size to ViewConfiguration
for use in semantics and pointer input to ensure accessiblity."

Test: new tests

Change-Id: Ie861ca1fcdbfcc9455352fc3a459d5734d5d57cc