Pin Compose Animation Dependencies to Compose 1.0.0
This is the first part in decoupling the Compose groups
so they can release independently.  The Runtime and Compiler
must remain tip-of-tree project dependencies because 1.0.0
is not compatible with Kotlin 1.5.21, so they were
not updated.

Additionally, the intra-group dependencies were kept as
project dependencies.

Updates to the latest/tip-of-tree versions should only
happen when needed.

Test: gw bOS
Change-Id: I5ed8efadf104df9dc720af732741046630aa3193