Fixed an issue that flash not working when setting flash mode shortly followed by taking picture.
On some devices, if set flash mode to ON and take pictures
immediately, it could cause AE precapture failed to work and thus
flash won't work.

Cause and solution:
It seems that if the repeating request for updating the AE mode
is not completed, triggering AE precapture will not work as
expected. The solution is that before trigger AE precapture,
ensure the repeating request for latest AE mode change is completed.

Relnote: "Fixed an issue that flash is not working when setting flash
mode to FLASH_MODE_ON shortly followed by taking pictures."
Test: Camera2CameraControlImplTest / Camera2CameraControlImplDeviceTest
Bug: 192916429

Change-Id: Ieb49b78a6b5bc9ea30b0edee91af76359454d9ac