MediaRouter: Correct group route ID in MR2
When a member route controller for a MR2 routing controller,
it finds the group route controller by route id.

Before this CL, RoutingController#getId was used to guess id but
it could be wrong if a group descriptor is provided by the provider.

It resulted in not being able to control member volume.

It also reverts change from aosp/1769465, which was wrong.

Bug: 194732610
Test: manually w/ and w/o supportsDynamicGroup in support v7 demos
Change-Id: I66dd3426e6db0cc30a6762f2d7743699bf3da6f4