API improvements for RemoteIntentHelper
Exposing action constant instead of helper functions.

Renaming class to be RemoteActivityHelper and functions
that are dealing with extras to have a prefix target
which makes it more useful.

Documenting meaning of the nodeId.

Bug: 194895941
Test: Existing ones
Relnote: "Renamed class RemoteIntentHelper to RemoteActivityHelper. Renamed functions getRemoteIntentExtraIntent and getRemoteIntentNodeId to getTargetIntent and getTargetNodeId, respectively."

Change-Id: Id2042393d23e7e7520e631e01f84bd40bff45ce3