Don't crash if ConfirmationActivity does not have a message.
aosp/1764469 introduced a bug that could lead to a null-dereference of
the overlay message (despite setMessage being @NonNull,
ConfirmationActivity could pass null here if the starting intent did not
contain a message). ConfirmationActivity will now check for a missing
message, and pass "" instead in this case.

Bug: 194985941
Test: New test added.
Relnote: "Fix bug that caused ConfirmationActivity to crash if no message
was provided."

Change-Id: Ie6055a81b7d1c696d4b4f7eb81ff03862f35f665