Add more info to project description in project-creator script
For library modules, the description should be updated to
read like "This library makes it easy for developers to...".

For non-library modules, this description is likely more
of a summary, such as samples or lint modules.

The script will now ask you for a description when
you create the project.

Fixes: b/186517077
Bug: b/181769383
Test: ./create_project.py androidx.foo foo-bar
Change-Id: I51478f3d094d08eb64f2a2c788c654d3825e78b3