Fix for the derived state read causing other state read to be forgotten
We do not immediately remove invalidated scopes in SnapshotStateObserver anymore. We don't need it as now we clear them right before rerunning the scope. One side effect is that onChanged lambda can now be re-executed again for the scope which was already invalidated, which is not a big issue given that all our scopes are handling that correctly.

Test: new tests in ModelReadsTest and SnapshotStateObserverTests
Fixes: 192595563
Change-Id: Ic04db3e30538eb72892b8c1b267b800d0c9b841a