Supporting non-POJO input in MapInfo.
Adding support for allowing primitive types as well as Strings in the keys and/or values of DAO methods with multimap return types. This is done via the use of the @MapInfo annotation. The developer is given the opportunity of providing the key and/or value column names that are needed to perform the provided query, and clarify using which column(s) the matching should be done.

In the follow-up CL, demonstrated code usage will be added to the @MapInfo annotation class, and additional warning message conditions for when @MapInfo is used redundantly or is not used when required will be added.

Test: QueryMethodProcessorTest.kt, MultimapQueryTest.java, RawQueryMethodProcessorTest.kt
Relnote: Supporting non-POJO keys/values for Multimap return types in DAOs.
Change-Id: I4d7048542b73b963fdb0b901dcab76f19ef3b5b8