Invoke unbundled perfetto via tracebox
Test: ./gradlew benchmark:benchmark-common:cC benchmark:benchmark-macro:cC
Fixes: 197144527

Switching to tracebox has several advantages:
 - no manual starting/killing of traced/traced_probes
 - smaller APK size (964Kb saved)
 - unifying bundled/unbundled perfetto codepaths

This does require a change of how we kill processes, as now the three
processes are all named "tracebox" (instead of
perfetto/traced/traced_probes), so we need to track PID from creation
time, and kill that. This also required that we kill preexisting
tracebox/perfetto processes in a new way.

This was always something we meant to do anyway, but has the side
benefit of moving a lot of logic out of PerfettoHelper.

Change-Id: I8d21b069e4e83224d55d8337a38a39551f268901