Update canClearCaptureRequestOptions test
1) Modify the request value from COLOR_CORRECTION_ABERRATION_MODE_OFF to
COLOR_CORRECTION_MODE_FAST.
2) The Camera template will add some extra keys into the CaptureRequest,
it would make it difficult to ensure the Keys are not in the request.
In the canClearCaptureRequestOptions test, we already verify the request
options that have been cleared by clearCaptureRequestOption() API, and
the camera has been updated, so we can skip the extra checking on the
captured result.
3) Pipe-integration will not verify the capture results from the CameraDevice,
it will verify the results from the Camera-pipe. So the verification in
the pipe-integration doesn't have the same problem like 2).

Test: Camera2CameraControlDeviceTest
Change-Id: Ie6d7920c5b7c9434fdc041e117416b110619f777