Generate layouts to handle dimension specification
Android S+ allows the developer to specify the layout_width and
layout_height properties, but this wasn't available before. Moreoever,
no version of Android (up to S at least) allow setting the layout weight
programatically for RemoteViews. All that is solved by resource
generation, creating all the needed combination of properties for each
view.

This CL generates layouts, and generate a mapping defining the content
of the various generated layouts.

It also adds a demo, showing how this works.

Test: ./gradlew :glance:glance-appwidget:connectedAndroidTest
Relnote: n/a
Bug: 199427087

Change-Id: I200380a61d31d30f4b8f98f47a2fd4fed359ac0c