Updates Material components to have a minimum touch target size
Button, Slider, Checkbox, RadioButton, and Switch have a visual size smaller than the minimum touch target size of 48x48dp. This means that even with the recent touch target expansion work, in many cases there will not be enough space for hit testing to expand, meaning that these components are not accessible. This CL adds extra space around these components if they are smaller than 48x48, to make sure that there will be enough size for the expanded hit testing to work, ensuring that these components have an accessible touch target.

Also updates Snackbar and AlertDialog to account for the new larger size of buttons in their layout implementation.

Bug: b/149691127
Fixes: b/171509422
Fixes: b/150377503
Fixes: b/174707736
Test: existing tests, new tests
Relnote: "Adds minimum touch target sizing for Material components that do not have an accessible touch target. As this adds extra spacing around components to make sure they have a large enough touch target, this may change existing UIs that assume the size of these components is their visual size, and doesn't account for touch target size. You can use the experimental LocalMinimumTouchTargetEnforcement composition local to disable this behavior across a hierarchy, but this is intended only to be a temporary escape hatch while updating existing UIs to account for the new minimum size."
Change-Id: I9b966a79a290538d8d450aa732b5092760cbece4