Introduce StorageException to datastore-core
The exception is defined as a subclass of IOException on Android and JRE.

The net effect of this change is only to make CorruptionException a
subclass of StorageException.

Note that:

1. The DataStore interface is still defined to be throwing IOException.
2. The catching sites in SingleProcessDataStore still expect
   IOException. They will need to be refactored later when the actual
   storage read/write logic gets extracted.
3. ReplaceFileCorruptionHandler continues to have the @Throws annotation
   that throws a checked IOException. This handler will be refactored
   later so that it will continue to have that annotation on Android and
   JRE.

Also note that none of the tests are changed and this is the intended
effect of this change. That is, all existing catching sites should not
be affected.

Relnote: CorruptionException now subclasses StorageException, but existing expectations (esp. catch sites that expect IOException) are not affected.
Test: Existing tests
Change-Id: Ia867f2dbc86dca503417d4d06c78d62f5c867bbb