Fix-up ScalingLazyListState performance scaling calculations
There are a couple of places in ScalingLazyListState where there is the potential for a lot of wasted .find() calls in the lazyListState.layoutInfo.visibleItemsInfo due to not correctly bounding ranges of item indexes being looked for. This patch calculates the min/max item Indexes to be considered when searching.

Test: androidx.wear.compose.material.ScalingLazyListLayoutInfoTest,androidx.wear.compose.material.ScalingLazyColumnTest
Bug: 199388253
Change-Id: Ib0b9b278dc6ca3770e7cc746a565f6fc1069dcce