Handle default java interface methods in inherited interfaces
Calling default interface methods in super is `weird` in java.

Previously, we used `Dao.super.method` to call these but that does not
work if `method` is inherited. e.g.
Dao -> BaseDao (and method is in BaseDao).

If the Dao is an abstract class (and BaseDao is an interface), we don't
have the necessary scope to call `Dao.super.method` (or
`BaseDao.super.method`). Instead, we need to call
`Dao_Impl.super.method`.

On the other hand, if `Dao` is an interface, we need to use
`Dao.super.method` as we have the necessary scope.

Added compilation tests with kotlin's different default method handlings
as well as. Unfortunately, we cannot run the compilation tests with KSP
because we don't know what compiler argument was passed.
https://github.com/google/ksp/issues/627

Also added a test case to Java TestApp (no compilation tests there as
java has only 1 way of handling this).

Fixes: 199546642
Test: DefaultsInDaoTest, DefaultDaoMethodsTest
Change-Id: I15f6dbe45b4d743daa84fce2016fcd4ed7f50fbc