Expand rememberNavController to add custom navigators
Navigation supports adding additional Navigators,
but did not have any official way of doing that in the
Navigiation Compose world. By expanding the API of
rememberNavController to allow adding additional
Navigators, we can ensure that developers have a
consistent way to solve this.

Relnote: "`rememberNavController()` now takes a optional
set of `Navigator` instances that will be added to the
returned `NavController` to better support optional
Navigators such as those from
[Accompanist Navigation Material](https://google.github.io/accompanist/navigation-material/)."
Test: Updated tests still pass
BUG: 193555369

Change-Id: I4619e6c4b47fc76c45a64c68085519fd8d18f699