Make MultiModalInjectionScope an interface
Which is implemented by MultiModalInjectionScopeImpl. The interface is
sealed to allow future addition of properties, and the only
implementation is internal so exhaustive when over the type of an
instance of that interface is forced to include an else block.

It will also improve layering of the API by preventing implementation
details to be leaked into the public API of MultiModalInjectionScope,
which would happen for example when it decides to implement Density by
delegation.

Bug: 190493367
Test: ./gradlew compose:ui:ui-test:cC && \
        ./gradlew compose:ui:ui-test:test
Relnote: N/A
Change-Id: Ib76c2ccbfe6393dc3397dfc7c3a0ec08ff4cd842