Improve Room's type deferring processing.
Use new APIs in Context to report missing types and thus further validate POJOs, DAOs and Type Converters. Database annotated elements are by default validated since those elements are the ones requested by the DatabaseProcessingStep. However, only superficial validation is done by xprocessing, it is up to the processor author to further validate the AST visited.

For Room this means calling a new validate() method provided in XElement that can be used to perform superficial validation on an element and further decided so defer processing or error out.

The main benefit of this change is that Room will now be able to better interact with other code generators, such as those that generated type converters.

Room will reject elements with missing types up to the last round, where it will try to continue processing as much as possible and generate code along with reporting errors including those os missing types.

Bug: 120045569
Test: ./gradlew room:room-compiler:test
Change-Id: I3f672cd568e46fc222325122d4a352c1567439d7