Add lint rule for attaching and detaching same fragment
Since Fragment 1.3.0 and the switch over to the new state manager, the
FragmentManager now optimizes all operations within a single
transactions. This means that doing detach() and attach() in the same
transactions without using `reorderingAllowed(true)` will no longer
destroy and recreate the fragment view.

This lint rule detects usages of detach() and attach() in the same
transaction and warns developers that if they would like the old
behavior they should separate the operations into two different
transactions.

RelNote: "Fragments will now notify you via lint if you are attempting
to `detach()` and `attach()` the same fragment in a single
`FragmentTransaction` that it is a no-op and you should separate those
operations into separate transactions."
Test: added AttachAndDetachInSameTransactionTest
Bug: 173472486
Bug: 200867930

Change-Id: I917af67b88125d844d389f637c169583b38c453c